migrate hook tests
I've moved these to test_plugin because although the tests
are calling actions we are really testing the behaviour of
our after_create/after_update/after_delete hooks
